removed unused classes. minor refactoring
