--- whois-commons/src/main/resources/nrtm_data.sql
@@ -1,2 +1,0 @@
-insert into source(id, name)
-values (100, 'RIPE_TEST_SOURCE');
