--- whois-nrtm4/src/main/java/net/ripe/db/nrtm4/SnapshotInformation.java
@@ -1,10 +1,0 @@
-package net.ripe.db.nrtm4;
-
-import java.util.Date;
-
-
-public class SnapshotInformation {
-    private Long whoisSerial;
-    private Long snapshotID;
-    private Date publishedDate;
-}
