--- whois-nrtm4/src/test/java/net/ripe/db/nrtm4/AbstractNrtm4IntegrationBase.java
@@ -1,8 +1,0 @@
-package net.ripe.db.nrtm4;
-
-import net.ripe.db.whois.common.dao.jdbc.AbstractDatabaseHelperIntegrationTest;
-
-
-public class AbstractNrtm4IntegrationBase extends AbstractDatabaseHelperIntegrationTest {
-
-}
